moved SongInfo classes from /view/song to /view
